Rutas admin, loggin mmiddleware, vistas añadidas

+++ routes/web.php
@@ -29,52 +29,56 @@ Route::get('/', [HomeController::class, 'getHome']);//Devuelve la vista del home
 /**
  * Control de usuarios
  */
-Route::get('/usuarios', [UserController::class, 'getUsers']); //Lista todos los usuarios
-Route::post('/usuarios', [UserController::class,'searchUsers']);//Buscador de usuarios por nombre o email
-Route::get('/usuarios/{id}', [UserController::class, 'getProfile']);
-Route::delete('/usuarios/{id}', [UserController::class, 'deleteUser']);
-Route::post('/usuarios/{id}', [UserController::class, 'editProfile']);
-
+Route::get('/admin/usuarios/', [UserController::class, 'getUsers'])->middleware(['auth']); //Lista todos los usuarios
+Route::post('/admin/usuarios/', [UserController::class,'searchUsers'])->middleware(['auth']);//Buscador de usuarios por nombre o email
+Route::get('/admin/usuarios/{id}', [UserController::class, 'getProfile'])->middleware(['auth']);
+Route::delete('/admin/usuarios/{id}', [UserController::class, 'deleteUser'])->middleware(['auth']);
+Route::post('/admin/usuarios/{id}', [UserController::class, 'editProfile'])->middleware(['auth']);
 
 
 /**
  * Control de Musculos
  */
-Route::get('/musculos', [MusculosController::class, 'getMusculos']); //Lista todos los usuarios
-Route::post('/musculos', [MusculosController::class, 'newMusculo']);
-Route::delete('/musculos/{id}',[MusculosController::class,'deleteMusculo']);
-
+Route::get('/admin/musculos', [MusculosController::class, 'getMusculos'])->middleware(['auth']); //Lista todos los usuarios
+Route::post('/admin/musculos', [MusculosController::class, 'newMusculo'])->middleware(['auth']);
+Route::delete('/admin/musculos/{id}',[MusculosController::class,'deleteMusculo'])->middleware(['auth']);
+Route::get('/admin/musculos/{id}', function() { #Evitar excepcion al acceder a esta pagina
+    return abort(404);
+});
 /**
  * Control de Ejercicios
  */
-Route::get('/ejercicios', [EjerciciosController::class, 'getEjercicios']); //Lista todos los ejercicios
-Route::post('/ejercicios', [EjerciciosController::class, 'newEjercicios']);//Añade ejercicios
-Route::delete('/ejercicios/{id}',[EjerciciosController::class,'deleteEjercicio']);//Elimina ejercicios
-Route::get('/ejercicios/{id}', [EjerciciosController::class, 'getEjercicioDetalle']);//Añade ejercicios
+Route::get('/admin/ejercicios', [EjerciciosController::class, 'getEjercicios'])->middleware(['auth']); //Lista todos los ejercicios
+Route::post('/admin/ejercicios', [EjerciciosController::class, 'newEjercicios'])->middleware(['auth']);//Añade ejercicios
+Route::delete('/admin/ejercicios/{id}',[EjerciciosController::class,'deleteEjercicio'])->middleware(['auth']);//Elimina ejercicios
+Route::get('/admin/ejercicios/{id}', [EjerciciosController::class, 'getEjercicioDetalle'])->middleware(['auth']);//Añade ejercicios
 
 
 /**
  * Control de gruposMusculares
  */
-Route::get('/gruposMusculares', [GrupoMuscularController::class, 'getGruposMusculares']); //Lista todos los grupos musculares
-Route::post('/gruposMusculares', [GrupoMuscularController::class, 'newGruposMusculares']);
-Route::delete('/gruposMusculares/{id}',[GrupoMuscularController::class,'deleteGrupoMuscular']);
-
+Route::get('/admin/gruposMusculares', [GrupoMuscularController::class, 'getGruposMusculares'])->middleware(['auth']); //Lista todos los grupos musculares
+Route::post('/admin/gruposMusculares', [GrupoMuscularController::class, 'newGruposMusculares'])->middleware(['auth']);
+Route::delete('/admin/gruposMusculares/{id}',[GrupoMuscularController::class,'deleteGrupoMuscular'])->middleware(['auth']);
+Route::get('/admin/gruposMusculares/{id}', function() { #Evitar excepcion al acceder a esta pagina
+    return abort(404);
+});
 
 /**
  * Control de Entrenamientos
  */
-Route::get('/entrenamientos', [EntrenamientosController::class, 'getEntrenamientos']);//Lista todos los entrenamientos
-Route::post('/entrenamientos', [EntrenamientosController::class, 'newEntrenamiento']);
-Route::get('/entrenamientos/{id}', [EntrenamientosController::class, 'getEntrenamientoDetalle']);
-Route::delete('/entrenamientos/{id}', [EntrenamientosController::class, 'deleteEntrenamiento']);
-Route::post('/entrenamientos/{id}', [EntrenamientosController::class, 'editEntrenamiento']);
+Route::get('/admin/entrenamientos', [EntrenamientosController::class, 'getEntrenamientos'])->middleware(['auth']);//Lista todos los entrenamientos
+Route::post('/admin/entrenamientos', [EntrenamientosController::class, 'newEntrenamiento'])->middleware(['auth']);
+Route::get('/admin/entrenamientos/{id}', [EntrenamientosController::class, 'getEntrenamientoDetalle'])->middleware(['auth']);
+Route::delete('/admin/entrenamientos/{id}', [EntrenamientosController::class, 'deleteEntrenamiento'])->middleware(['auth']);
+Route::post('/admin/entrenamientos/{id}', [EntrenamientosController::class, 'editEntrenamiento'])->middleware(['auth']);
+
 
 /**
  * Controlador del home
  */
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'getHome'])->name('home');
 
-Route::get('/admin', [AdministracionController::class, 'getAdministracion']);//Lista todos los entrenamientos
+Route::get('/admin', [AdministracionController::class, 'getAdministracion'])->middleware(['auth']);//Lista todos los entrenamientos
 
 Auth::routes();
